Poison damage system complete
both the player and enemies can gain the poison status by being it by an attack that has the status. (25% chance)
the status clears back to normal if the target's health is below 25% of the maximum. the health bar of the target changes color if poisoned and will go back to the original color